+++ core/src/main/scala/geotrellis/store/query/Repository.scala
@@ -16,7 +16,7 @@
 
 package geotrellis.store.query
 
-trait QueryCollection[T, G[_]] {
-  val list: G[T]
+trait Repository[T, G[_]] {
+  def store: G[T]
   def find(query: Query): G[T]
 }
