+++ core/src/main/scala/geotrellis/store/query/package.scala
@@ -29,16 +29,6 @@ package object query {
     def and(r: Query): Query = QueryF.and(self, r)
   }
 
-  implicit class RasterSourceOps(self: RasterSource) {
-    def projectedExtent: ProjectedExtent         = ProjectedExtent(self.extent, self.crs)
-  }
-
-  implicit class ProjectedExtentOps(self: ProjectedExtent) {
-    def intersects(pe: ProjectedExtent): Boolean = self.extent.intersects(pe.reproject(self.crs))
-    def covers(pe: ProjectedExtent): Boolean     = self.extent.covers(pe.reproject(self.crs))
-    def contains(pe: ProjectedExtent): Boolean   = self.extent.contains(pe.reproject(self.crs))
-  }
-
   def or(l: Query, r: Query): Query          = QueryF.or(l, r)
   def and(l: Query, r: Query): Query         = QueryF.and(l, r)
   def nothing: Query                         = QueryF.nothing
@@ -50,4 +40,14 @@ package object query {
   def covers(pe: ProjectedExtent): Query     = QueryF.covers(pe)
   def at(t: ZonedDateTime, fieldName: Symbol = 'time): Query                          = QueryF.at(t, fieldName)
   def between(t1: ZonedDateTime, t2: ZonedDateTime, fieldName: Symbol = 'time): Query = QueryF.between(t1, t2, fieldName)
+
+  implicit class RasterSourceOps(self: RasterSource) {
+    def projectedExtent: ProjectedExtent = ProjectedExtent(self.extent, self.crs)
+  }
+
+  implicit class ProjectedExtentOps(self: ProjectedExtent) {
+    def intersects(pe: ProjectedExtent): Boolean = self.extent.intersects(pe.reproject(self.crs))
+    def covers(pe: ProjectedExtent): Boolean     = self.extent.covers(pe.reproject(self.crs))
+    def contains(pe: ProjectedExtent): Boolean   = self.extent.contains(pe.reproject(self.crs))
+  }
 }
