+++ ogc/src/main/scala/geotrellis/server/ogc/wmts/CapabilitiesView.scala
@@ -209,6 +209,6 @@ object CapabilitiesView {
   def modelAsLayers(wmtsModel: WmtsModel): List[DataRecord[LayerType]] =
     wmtsModel
       .sources
-      .list
+      .store
       .map { src => DataRecord("wms".some, "Layer".some, src.toLayerType(wmtsModel.matrices)) }
 }
